# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AGS Documentation",
        "description": "Cette API permet de gérer une plateforme complète de commerce en ligne. Elle prend en charge les opérations suivantes Inscription et connexion des utilisateurs : Gestion des comptes utilisateurs avec des fonctionnalités CRUD (Créer, Lire, Mettre à jour, Supprimer) Gestion des partenaires : CRUD complet pour les partenaires associés à la plateforme. Commentaires : Ajout et gestion des commentaires des utilisateurs Gestion du panier : Ajout et suppression d'articles, ainsi qu'une gestion complète du panier d'achat. Commandes : Création et suivi des commandes. Achats de produits et de formations : Intégration des services d'achat pour divers produits et formations Données de test des utilisateurs Administrateur : Email : admin@example.com Mot de passe : adminpassword Étudiant : Email : etudiant@example.com Mot de passe : etudiantpassword Client : Email : client@example.com Mot de passe : clientpassword",
        "version": "1.0.0"
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "consumes": [
        "multipart/form-data"
    ],
    "paths": {
...
        "/api/logout": {
            "post": {
                "summary": "déconnexion",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "summary": "rafrechir",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {}
                            },
                            "example": {}
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "connexion",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "email",
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "password",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            },
                            "example": {
                                "email": "admin@example.com",
                                "password": "adminpassword"
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "inscription",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "Authentification"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nom_complet": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            },
                            "example": {
                                "nom_complet": "mouctardiakhate",
                                "email": "moctar.diakhate@exemple.com",
                                "telephone": "778552525",
                                "password": "123456",
                                "photo": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}